drm/i915: Create a more generic pm handler for hsw+
HSW has some special requirements for the VEBOX. Splitting out the
interrupt handler will make the code a bit nicer and less error prone
when we begin to handle those.

The slight functional change in this patch (queueing work while holding
the spinlock) is intentional as it makes a subsequent patch a bit nicer.
The change should also only effect HSW platforms.

Based on patches from:
CC: Haihao Xiang <haihao.xiang@intel.com>
Signed-off-by: Ben Widawsky <ben@bwidawsk.net>