drm/i915: HSW PTE encoding for ppgtt
After the previous patch to extract pte encoding, it now becomes more
useful to pass in the cache level to do the PTE encoding.  There are two
primary reasons:
1. Address encoding differs by platform
2. Haswell will add some new attributes, and this is a preliminary step
for supporting that.

This also serves as the ppgtt analog of
<fill in sha when done>
drm/i915: Some HSW PTE fixes

since it does indeed fix the PTE encoding in PPGTT.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>