drm/i915: Add new HSW cache attributes
This is split out for easy review. There should be no functionality
changes here.

Also included in the patch is the encoding. It has no consumer yet, but
this patch made it much easier to review, and it will be used very soon.

Haswell uses bit 11, and 3:1 for the cache attributes. There isn't
really a great pattern defining the encoding.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>