drm/i915: add policy and age to plumbing
There is no 1:1 correspondence with the old policy known only as "LLC"
so we call that Writeback with age 3. As long as all PTEs contain the
same eviction policy, the results should be similar to previous
generations.

Notice that we've not added AGP types yet, and we've not modified the
pte encoding in ppgtt_bind yet. As such, this shouldn't actually impact
anything yet.

Also note, I use age 0 for the scratch pages. There isn't a terribly
great reason for doing this other than differentiating from in use
objects because these pages should never end up in the cache anyway.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>