drm/i915: Support new cache bits
With the plumbing complete, all we need to do is actually encode the
ppgtt ptes now, and deal with the archaic AGP types. To solve this
problem, since we already have a nice function for encoding ptes is we
create a new AGP flag which tells the AGP layer that we've already
encoded the pte.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>