drm/i915: displayable surfaces writethrough (HSW)
Docs state that when LLC and eLLC both support Write-Through memory
type, we must mark any displayable surfaces as WT cacheable.

[REMOVE BEFORE UPSTREAM]
Note that the docs say we need this on DevHSW:GT3e. Since we're not
exactly sure what that means, I've taken a guess here. We need to verify
this when possible.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>