drm/i915: Use eLLC by default when available
We can conceptually think of eLLC as like an L2 of the shared cache.
Assuming that's a correct way to conceptualize it, I think a good
default is to have the LLC backed by eLLC.

The doc does speak of certain special objects which don't belong in
either LLC or eLLC based on the way they are used/aged, but that
optimization is left up to userspace on the interface we've defined
previously for controlling the cache knobs.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>