mei: add mei_stop function to stop mei device
mei_stop calls mei_reset with disabling the interrupts.
It will have the same effect as the open code it replaces in the mei_remove.

The reset sequence on remove is required for the Lynx Point LP devices
to clean the reset state.

mei_stop is called from mei_pci_suspend and mei_remove functions

Change-Id: If32d4238a3e6e73398b24b55a5c52569061c8e16
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>