mei: reset client connection state on timeout
On connection timeout we leave the connecting client in
connecting state. Since a new connection is stalled till
previous connection is completed in this case no new connection
is possible till the user space does release the file handle.
Therefore on timeout we move the client to disconnected state.

Stable: #3.15
Backport: #3.10-3.14
Change-Id: I39eee4408ff7b9d0361122b9e714e7dd237ce16a
Signed-off-by: Alexander Usyskin <alexander.usyskin@intel.com>
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>