UPSTREAM: ASoC: rt5616: add the mclk for the codec driver
This patch adds the code to enable the clock to the CODEC driver
if it needs the clock enabled.

In some case, We need to claim the clock which is driving the codec
so that when we enable clock gating, we continue to clock the codec
when needed.

if mclk provided, to enable and disable the clock source.

BUG=chrome-os-partner:50756
TEST=Build and boot on rk3288

Change-Id: I1b551f9d25a2e5c4296214733fccdf561a4238df
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 76d3204eaa9364a662f0fe0f075dfe61e6bd14fe)